fix: padding of tensors in a batch
Previously when collating the tensors in a batch,
they were padded from the right but this hindered
the ability of the model to learn.

Now, when collating the tensor, they are padded from
the left. We get a much more better result by doing
this because the rnn reads a word from left to right,
and when padding from the left, the rnn doesn't start
learning until the actual characters (tensors representing
characters) start.