refactor: dataframe to tuple
Instead of using a pandas dataframe, a list of tuple
is used in the custom dataset.

The get_names function is refactored to return the
list of tuples only as the categories are already
provided in the metadata.

The create_dataframe function is now removed as it
has no use right now.